feat: making aldev work on Windows

+++ templates/tilt.go
@@ -12,7 +12,6 @@ config.define_bool('api-only', usage='use this to only work with the API part of
 cfg = config.parse()
 useLocalDeps = cfg.get('use-local')
 apiOnly = cfg.get('api-only')
-devMode = "--api" if apiOnly else "--web"
 
 # describing the deployment of all the backend services, and their configuration
 if useLocalDeps or apiOnly:
@@ -27,7 +26,7 @@ else:
 # building the API's code
 local_resource(
     name  ='{{.AppName}}-api-compile',
-    cmd   ='aldev codegen '+devMode,
+    cmd   ='aldev codegen',
     # taking into account the dependencies
     deps  =['{{.API.SrcDir}}', '../goald', '../emerald'], # REMOVE EMERALD
     # the API config is also ignored here, because Aldev is already watching it
@@ -51,37 +50,58 @@ docker_build_with_restart(
 # deploying the API
 k8s_resource('{{.AppName}}-api-depl', resource_deps=['{{.AppName}}-api-compile'])
 
+# --- PROXY part (for WSL) ----------------------------------------------------
+
 # getting the load balancer's IP - cf https://docs.tilt.dev/extensions.html
 apiHost = str(local(echo_off=True, command="kubectl get services --namespace kube-system "+
   "-o jsonpath='{.items[?(@.spec.type==\"LoadBalancer\")].status.loadBalancer.ingress[0].ip}'"))
 
 # since the load balancer is on host 'apiHost' and not 'localhost', we can reach it by sending calls
 # to localhost:apiPort - from Windows for instance - and proxying them inside Linux to apiHost:apiPort
-local_resource(
-    name         ='{{.AppName}}-lb-proxy',
-    cmd          ='killall -q socat || true',
-    serve_cmd    ='socat TCP-LISTEN:{{.API.Port}},fork TCP:'+apiHost+':{{.API.Port}}',
-    resource_deps=['{{.AppName}}-api-depl'],
-    )
+if os.name != 'nt':
+  local_resource(
+      name         ='{{.AppName}}-lb-proxy',
+      cmd          ='killall -q socat || true',
+      serve_cmd    ='socat TCP-LISTEN:{{.API.Port}},fork TCP:'+apiHost+':{{.API.Port}}',
+      resource_deps=['{{.AppName}}-api-depl'],
+      )
 
 # --- WEB part ----------------------------------------------------------------
+
+installCmd = 'rm -fr node_modules/.vite; npm i --force'
+if os.name == 'nt':
+  installCmd = 'rmdir /s /q node_modules/.vite 2>nul || echo "no Vite cache for now" && npm i --force' # Windows equivalent
+
+webDep = '{{.AppName}}-lb-proxy'
+if os.name == 'nt':
+  webDep = '{{.AppName}}-api-depl'
 
 if not apiOnly:
   # if we're developing libraries along with this project, then:
   if useLocalDeps:
     # these are the env vars defined in the aldev config file (.aldev.yaml)
     webAppEnvVars = "WEB_API_URL=http://localhost:{{.API.Port}}"
-    {{range .Web.EnvVars}}webAppEnvVars += " {{.Name}}={{.Value}}"
+    if os.name == 'nt':
+      webAppEnvVars = "set WEB_API_URL=http://localhost:{{.API.Port}} &&"
+
+    envVarsSuffix = ''
+    envVarsSep = ''
+    if os.name == 'nt':
+      envVarsSuffix = ' &&'
+      envVarsPrefix = ' set'
+
+    {{range .Web.EnvVars}}webAppEnvVars += envVarsPrefix + " {{.Name}}={{.Value}}" + envVarsSuffix
     {{end}}
+
     # locally running Vite's dev server - no need to containerize this for now
     local_resource(
         name         ='{{.AppName}}-vite-serve',
         dir          ='{{.Web.SrcDir}}',
-        cmd          ='rm -fr node_modules/.vite && npm i --force',
+        cmd          =installCmd,
         deps         =['{{.Web.SrcDir}}/package.json'],
         serve_cmd    =webAppEnvVars + ' npm run dev',
         serve_dir    ='{{.Web.SrcDir}}',
-        resource_deps=['{{.AppName}}-lb-proxy'],
+        resource_deps=[webDep],
         )
   else:
     docker_build(
